backend/requirements.txt
Django==5.1.6
django-cors-headers==4.6.0
oracledb==2.5.1
PyJWT==2.10.1
python-dotenv==1.0.1
gunicorn==23.0.0

